Typings and reorg for initializing store and dispatchers
