+++ src/dispatchers/types.ts
@@ -1,0 +1,4 @@
+export interface WebsocketDispatcherProps {
+  port: number;
+  clientId: string;
+}
